requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
Django==5.1.6
django-crispy-forms==2.3
django-htmx==1.22.0
idna==3.10
pillow==11.1.0
python-decouple==3.8
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.3
urllib3==2.3.0
